# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMPARO - Sistema de Gestão de Casas de Repouso
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List daily activities
      parameters:
        - name: residentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [completed, pending, not_done]
      responses:
        "200":
          description: Activities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Activity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: residentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, completed, cancelled]
      responses:
        "200":
          description: Appointments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Schedule appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel/delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /handovers:
    get:
      operationId: listHandovers
      tags: [handover]
      summary: List shift handovers
      responses:
        "200":
          description: Handovers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Handover"
    post:
      operationId: createHandover
      tags: [handover]
      summary: Create shift handover record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoverInput"
      responses:
        "201":
          description: Handover created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handover"
...
  /handovers/{id}/confirm:
    post:
      operationId: confirmHandoverRead
      tags: [handover]
      summary: Confirm reading of handover
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handover"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all alerts
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [medication, activity, appointment, incident]
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Mark alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [caregiver, nurse, manager]
        shift:
          type: ["string", "null"]
          enum: [morning, afternoon, night, null]
        createdAt:
          type: string
          format: date-time
...
    Activity:
      type: object
      required: [id, residentId, type, date, status]
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [bath, feeding, hygiene, physiotherapy, medication, other]
        date:
          type: string
          format: date
        time:
          type: ["string", "null"]
        status:
          type: string
          enum: [completed, pending, not_done]
        observations:
          type: ["string", "null"]
        shift:
          type: string
          enum: [morning, afternoon, night]
        createdAt:
          type: string
          format: date-time
    ActivityInput:
      type: object
      required: [residentId, type, date, shift]
      properties:
        residentId:
          type: integer
        type:
          type: string
          enum: [bath, feeding, hygiene, physiotherapy, medication, other]
        date:
          type: string
          format: date
        time:
          type: string
        shift:
          type: string
          enum: [morning, afternoon, night]
        observations:
          type: string
    ActivityUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [completed, pending, not_done]
        time:
          type: string
        observations:
          type: string
    Appointment:
      type: object
      required: [id, residentId, date, time, specialty, status]
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        time:
          type: string
        specialty:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        observations:
          type: ["string", "null"]
        cancellationReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [residentId, date, time, specialty]
      properties:
        residentId:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
        specialty:
          type: string
        observations:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        date:
          type: string
        time:
          type: string
        specialty:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        observations:
          type: string
        cancellationReason:
          type: string
    Handover:
      type: object
      required: [id, shift, date, createdBy, occurrences, isRead]
      properties:
        id:
          type: integer
        shift:
          type: string
          enum: [morning, afternoon, night]
        date:
          type: string
          format: date
        createdBy:
          type: string
        occurrences:
          type: string
        pendencies:
          type: ["string", "null"]
        isRead:
          type: boolean
        readBy:
          type: ["string", "null"]
        readAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    HandoverInput:
      type: object
      required: [shift, occurrences]
      properties:
        shift:
          type: string
          enum: [morning, afternoon, night]
        occurrences:
          type: string
        pendencies:
          type: string
    Alert:
      type: object
      required: [id, type, title, message, severity, resolved, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [medication, activity, appointment, incident]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        resolved:
          type: boolean
        residentId:
          type: ["integer", "null"]
        residentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
...