# lib/api-zod/src/generated/types/handoverInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AMPARO - Sistema de Gestão de Casas de Repouso
 * OpenAPI spec version: 0.1.0
 */
import type { HandoverInputShift } from './handoverInputShift';

export interface HandoverInput {
  shift: HandoverInputShift;
  occurrences: string;
  pendencies?: string;
}