# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AMPARO - Sistema de Gestão de Casas de Repouso
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserShift } from './userShift';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  shift?: UserShift;
  createdAt?: Date;
}